More casts; bump upstream version; fix metadata

+++ crypto_sign/dilithium3/clean/poly.c
@@ -546,8 +546,8 @@ void PQCLEAN_DILITHIUM3_CLEAN_polyeta_pack(uint8_t *r, const poly *a) {
     DBENCH_START();
 
     for (i = 0; i < N / 2; ++i) {
-        t[0] = ETA - a->coeffs[2 * i + 0];
-        t[1] = ETA - a->coeffs[2 * i + 1];
+        t[0] = (uint8_t) (ETA - a->coeffs[2 * i + 0]);
+        t[1] = (uint8_t) (ETA - a->coeffs[2 * i + 1]);
         r[i] = t[0] | (t[1] << 4);
     }
 
@@ -754,12 +754,12 @@ void PQCLEAN_DILITHIUM3_CLEAN_polyz_pack(uint8_t *r, const poly *a) {
         t[0] = GAMMA1 - a->coeffs[2 * i + 0];
         t[1] = GAMMA1 - a->coeffs[2 * i + 1];
 
-        r[5 * i + 0]  = t[0];
-        r[5 * i + 1]  = t[0] >> 8;
-        r[5 * i + 2]  = t[0] >> 16;
-        r[5 * i + 2] |= t[1] << 4;
-        r[5 * i + 3]  = t[1] >> 4;
-        r[5 * i + 4]  = t[1] >> 12;
+        r[5 * i + 0]  = (uint8_t) t[0];
+        r[5 * i + 1]  = (uint8_t) (t[0] >> 8);
+        r[5 * i + 2]  = (uint8_t) (t[0] >> 16);
+        r[5 * i + 2] |= (uint8_t) (t[1] << 4);
+        r[5 * i + 3]  = (uint8_t) (t[1] >> 4);
+        r[5 * i + 4]  = (uint8_t) (t[1] >> 12);
     }
 
     DBENCH_STOP(*tpack);
@@ -811,7 +811,7 @@ void PQCLEAN_DILITHIUM3_CLEAN_polyw1_pack(uint8_t *r, const poly *a) {
     DBENCH_START();
 
     for (i = 0; i < N / 2; ++i) {
-        r[i] = a->coeffs[2 * i + 0] | (a->coeffs[2 * i + 1] << 4);
+        r[i] = (uint8_t) (a->coeffs[2 * i + 0] | (a->coeffs[2 * i + 1] << 4));
     }
 
     DBENCH_STOP(*tpack);
